pyproject.toml
[tool.poetry]
name = "aitubervts"
version = "0.1.0"
description = ""
authors = ["nahrun1682 <kawamura100995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.14"
websockets = "^15.0.1"
openai = "^1.93.0"
langchain = "^0.3.26"
httpx = "^0.28.1"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
